requirements.txt
# Web Framework
streamlit==1.32.0

# Data Processing
pandas==2.2.0

# Web Scraping & API Clients
apify-client==1.7.0
beautifulsoup4==4.12.3
requests==2.31.0
praw==7.7.1

# Database (MongoDB Atlas)
pymongo==4.6.0
dnspython==2.4.2

# AI & LLM (OpenRouter via OpenAI client)
openai==1.12.0
httpx==0.26.0

# Environment Variables
python-dotenv==1.0.1